Remove dead utils, move utils used only once
